+++ tests/Webtrees/MatchSeederTest.php
@@ -69,13 +69,13 @@ final class MatchSeederTest extends TestCase
      */
     protected function tearDown(): void
     {
-        $files = glob($this->dir . '/*');
+        if (($this->dir !== '') && is_dir($this->dir)) {
+            $files = glob($this->dir . '/*');
 
-        if ($files !== false) {
-            array_map('unlink', $files);
-        }
+            if ($files !== false) {
+                array_map('unlink', $files);
+            }
 
-        if (is_dir($this->dir)) {
             rmdir($this->dir);
         }
